refactor(schemas): the query-parameter vocabulary is not an MCP concept (#10061)
`netuidSchema`, `limitSchema`, `offsetSchema`, `ss58Schema`, `windowSchema`,
`sortSchema` and thirteen siblings lived in `schemas-src/mcp-tools/shared.ts`,
where the REST route schemas could not reach them -- so `schemas-src/routes/`
hand-declared the same parameters instead, as unbounded stubs.

Measured before this move: of 82 MCP input schemas that share properties with
their route's query schema, 56 disagree about the CONSTRAINTS, and the route
side is the looser one every time, because nothing has ever read it:

  netuid        MCP max 65535      route max 9007199254740991 (z.int(), no bound)
  offset        MCP max 1000000    route unbounded
  window        MCP enum(7d..all)  route z.string()
  counterparty  MCP ss58 pattern   route z.string()

That is what two declarations of one parameter always becomes. This move does
not fix those 56 -- that is #10062 -- it removes the reason they exist.

Nothing is published differently. `mcp-tools/shared.ts` re-exports every symbol,
so the ~100 tool modules importing them are untouched, and both proofs hold:
all 227 emitted MCP inputSchemas byte-identical, openapi.json byte-identical.

First of #10060.

Closes #10061
